Branch Created
Created my own branch to help with overwriting things.

I have created the blade for the Insert page, as well as the Model and Controller. I had to make a new Migration for Items.

The Controller can be used for the Home and Edit pages as well.

Low Stock Reminder function can be done once the Home page is finished.