Location and Name Integration
I have updated the Edit, Create, pages to use dropdowns and utilise the Name and Location tables.

Models for both added.

I added the framework to send Location and Name IDs to the homepage.

I also believe I fixed the error with edit and create. (They were re-directing to homepage, not items.homepage)

You will have to run a migration to drop the ItemName and Location columns from the Items table, since that is now covered by the foreign keys